3rd-assignment/requirements.txt
click==8.1.3
Flask==2.2.2
flask-cors==3.0.10  
flask-sqlalchemy==3.0.2
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
mysql-connector-python==8.0.32
packaging==22.0
SQLAlchemy==1.4.46
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0